Final version of the InputManager
This is (theoretically, I hope) the last update for the InputManager.
Everything has been cleaned up and is now fully functional.

The only known issue is a harmless warning that occurs when unplugging a joystick during gameplay. I plan to address this later if time allows, as a matter of good practice.